requirements.txt
Django>=5.1,<5.2
wagtail>=6.4,<6.5
psycopg2
python-dotenv